:sparkles: Convert revert to draft button to a component.
Add ability to hide itself when the travel authorization is not revertible.
